CORE-9384 Classes moved to "verifier" package
