+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/transaction/verifier/UtxoLedgerTransactionVerifierComponent.kt
@@ -1,12 +1,12 @@
-package net.corda.ledger.utxo.flow.impl.verification
+package net.corda.ledger.utxo.flow.impl.transaction.verifier
 
 import net.corda.flow.external.events.executor.ExternalEventExecutor
 import net.corda.ledger.utxo.contract.verification.CordaPackageSummary
 import net.corda.ledger.utxo.data.transaction.ContractVerificationStatus
 import net.corda.ledger.utxo.data.transaction.UtxoLedgerTransactionContainer
 import net.corda.ledger.utxo.data.transaction.UtxoLedgerTransactionInternal
-import net.corda.ledger.utxo.flow.impl.verification.events.VerifyContractsExternalEventFactory
-import net.corda.ledger.utxo.flow.impl.verification.events.VerifyContractsParameters
+import net.corda.ledger.utxo.flow.impl.transaction.verifier.external.events.VerifyContractsExternalEventFactory
+import net.corda.ledger.utxo.flow.impl.transaction.verifier.external.events.VerifyContractsParameters
 import net.corda.sandbox.type.SandboxConstants.CORDA_SYSTEM_SERVICE
 import net.corda.sandbox.type.UsedByFlow
 import net.corda.v5.application.serialization.SerializationService
@@ -20,20 +20,21 @@ import org.osgi.service.component.annotations.Reference
 import org.osgi.service.component.annotations.ServiceScope.PROTOTYPE
 import java.nio.ByteBuffer
 
+// TODO Rename this class to UtxoLedgerTransactionVerifier (CORE-9385)
 @Component(
-    service = [ UtxoLedgerVerificationService::class, UsedByFlow::class ],
+    service = [ UtxoLedgerTransactionVerifierComponent::class, UsedByFlow::class ],
     property = [ CORDA_SYSTEM_SERVICE ],
     scope = PROTOTYPE
 )
-class UtxoLedgerVerificationServiceImpl @Activate constructor(
+class UtxoLedgerTransactionVerifierComponent @Activate constructor(
     @Reference(service = ExternalEventExecutor::class)
     private val externalEventExecutor: ExternalEventExecutor,
     @Reference(service = SerializationService::class)
     private val serializationService: SerializationService
-) : UtxoLedgerVerificationService, UsedByFlow, SingletonSerializeAsToken {
+) : UsedByFlow, SingletonSerializeAsToken {
 
     @Suspendable
-    override fun verifyContracts(transaction: UtxoLedgerTransaction) {
+    fun verifyContracts(transaction: UtxoLedgerTransaction) {
         val verificationResult = externalEventExecutor.execute(
             VerifyContractsExternalEventFactory::class.java,
             VerifyContractsParameters(
