--- components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/verification/UtxoLedgerVerificationService.kt
@@ -1,20 +1,0 @@
-package net.corda.ledger.utxo.flow.impl.verification
-
-import net.corda.v5.base.annotations.Suspendable
-import net.corda.v5.ledger.utxo.ContractVerificationException
-import net.corda.v5.ledger.utxo.transaction.UtxoLedgerTransaction
-
-/**
- * [UtxoLedgerVerificationService] verifies contracts of UTXO ledger transactions.
- */
-interface UtxoLedgerVerificationService {
-    /**
-     * Verifies contracts of [UtxoLedgerTransaction].
-     *
-     * @param transaction UTXO ledger transaction.
-     *
-     * @throws [ContractVerificationException] if verification was unsuccessful.
-     */
-    @Suspendable
-    fun verifyContracts(transaction: UtxoLedgerTransaction)
-}
