+++ components/ledger/ledger-utxo-flow/src/test/kotlin/net/corda/ledger/utxo/flow/impl/transaction/verifier/external/events/VerifyContractsExternalEventFactoryTest.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.utxo.flow.impl.verification.external.events
+package net.corda.ledger.utxo.flow.impl.transaction.verifier.external.events
 
 import net.corda.data.KeyValuePairList
 import net.corda.data.flow.event.external.ExternalEventContext
@@ -6,8 +6,6 @@ import net.corda.flow.state.FlowCheckpoint
 import net.corda.ledger.utxo.contract.verification.CordaPackageSummary
 import net.corda.ledger.utxo.contract.verification.VerifyContractsRequest
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.ALICE_X500_HOLDING_IDENTITY
-import net.corda.ledger.utxo.flow.impl.verification.events.VerifyContractsExternalEventFactory
-import net.corda.ledger.utxo.flow.impl.verification.events.VerifyContractsParameters
 import net.corda.schema.Schemas
 import net.corda.virtualnode.toCorda
 import org.junit.jupiter.api.Assertions.assertEquals
